requirements.txt
Abg==2.3.1
git+https://github.com/KurimuzonAkuma/pyrogram@dev
pymongo
motor
requests
dnspython
tgcrypto==1.2.2
asyncio
python-dotenv
pycgi
apscheduler
Flask
uvloop
psutil
speedtest-cli
pyrogram==1.4.16
